STM-22 Moved the downloader runner to the downloader package. Moved the AlphaVantage data classes under the same alphavantage package as AlphaVantageDownloader.
