+++ src/main/java/com/robertozagni/SPYTM/data/collector/SpytmDataCollectorApplication.java
@@ -1,7 +1,7 @@
 package com.robertozagni.SPYTM.data.collector;
 
+import com.robertozagni.SPYTM.data.collector.downloader.StockDataDownloaderRunner;
 import com.robertozagni.SPYTM.data.collector.downloader.alphavantage.AlphaVantageDownloader;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
