+++ src/main/java/com/robertozagni/SPYTM/data/collector/downloader/alphavantage/AlphaVantageDownloader.java
@@ -1,9 +1,7 @@
 package com.robertozagni.SPYTM.data.collector.downloader.alphavantage;
 
 import com.robertozagni.SPYTM.data.collector.model.*;
-import com.robertozagni.SPYTM.data.collector.model.alphavantage.AVTimeSerie;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
